+++ CHANGELOG.md
@@ -52,6 +52,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - Future work will add other collection options which will pad to the full block duration
 - Add page to reorder channels (edit channel numbers) using drag and drop
   - New page is at **Channels** > **Edit Channel Numbers**
+- Scripted schedules: add setting to configure timeout of scripted playout build
+  - New setting is at **Settings** > **Playout** > **Scripted Schedule Timeout**
 
 ### Fixed
 - Fix green output when libplacebo tonemapping is used with NVIDIA acceleration and 10-bit output in FFmpeg Profile